pyproject.toml
[tool.poetry]
name = "bill-faster"
version = "0.1.0"
description = ""
authors = ["Le Xuan Manh <lexuanmanh101199@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.1"
python-dotenv = "^1.0.1"
uvicorn = "^0.30.6"
aiohttp = "^3.10.5"
python-multipart = "^0.0.9"
python-jose = "^3.3.0"
asyncpg = "^0.29.0"
passlib = "^1.7.4"
fastapi-babel = "^0.0.9"
boto3 = "^1.35.1"
fastapi-utils = "^0.7.0"
httptools = "^0.6.1"
httpx = "^0.27.2"
black = "^25.1.0"
pre-commit = "^4.1.0"
bcrypt = "^4.2.1"
greenlet = "^3.1.1"
jinja2 = "^3.1.5"
typing-inspect = "^0.9.0"
redisearch = "^2.1.1"
setuptools = "^75.8.2"
sqlalchemy = "^2.0.39"
psycopg2 = "^2.9.10"
psycopg2-binary = "^2.9.10"
alembic = "^1.15.1"
authlib = "^1.6.0"
google-auth = "^2.40.3"
itsdangerous = "^2.2.0"
requests = "^2.32.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
